Begin implementation of the REST API

+++ database/migrations/2018_12_09_022901_create_projects_table.php
@@ -14,13 +14,13 @@ class CreateProjectsTable extends Migration
     public function up()
     {
         Schema::create('projects', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->uuid('uuid')->unique();
             $table->string('name');
             $table->text('description');
             $table->boolean('default_project');
             $table->boolean('is_active');
-            $table->unsignedInteger('owner_id');
+            $table->unsignedBigInteger('owner_id');
             $table->timestamps();
             $table->foreign('owner_id')->references('id')->on('users')->onDelete('cascade');
             $table->unique(['name', 'owner_id']);
